requirements.txt
Flask==2.0.1
pyaml==21.8.3
requests==2.26.0